User Registration

+++ base/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.http import HttpResponse
@@ -63,7 +64,11 @@ def room(request, pk):
   context = {'room': room}
   return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
+  
+  if request.user != room.host:
+    return HttpResponse('You are not allowed here!!')
   form = RoomForm()
   if request.method == 'POST':
     form = RoomForm(request.POST)
@@ -75,23 +80,28 @@ def createRoom(request):
   context = {'form': form}
   return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
-  
-  room = Room.objects.get(id=pk)
-  form = RoomForm(instance=room)
-  
-  if request.method == 'POST':
-    form = RoomForm(request.POST, instance=room)
-    if form.is_valid():
-      form.save()
-      return redirect('home')
-  
-  context = {'form': form}
-  
-  return render(request, 'base/room_form.html', context)
+    room = Room.objects.get(id=pk)  # Fetch the room object first
 
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
+
+    form = RoomForm(instance=room)
+
+    if request.method == 'POST':
+        form = RoomForm(request.POST, instance=room)
+        if form.is_valid():
+            form.save()
+            return redirect('home')
+
+    context = {'form': form}
+    return render(request, 'base/room_form.html', context)
+
+@login_required(login_url='login')
 def deleteRoom(request, pk):
+  if request.user != room.host:
+    return HttpResponse('You are not allowed here!!')
   room = Room.objects.get(id=pk)
   if request.method == 'POST':
     room.delete()
